Fix cobrand restriction of My/Nearby.
5c79337 simplified a bit too far, as Nearby and My could only then
in a particular cobrand potentially restrict to a particular body,
not any other criteria. Switch to more generic updates_restriction
and problems_restriction to allow cobrands more flexibility.

Make sure a few tests delete their bodies fully so that new tests
pass when run as part of the suite.

Fixes #1289.